+++ expo/app/(authenticated)/index.tsx
@@ -1,42 +1,12 @@
-import { Button, TextInput } from "react-native";
+import { Button } from "react-native";
 
 import { useSession } from "@/ctx";
 import useAppAxios from "@/hooks/useAppAxios";
-import { useState } from "react";
 import { ThemedView } from "@/components/ThemedView";
 import { ThemedText } from "@/components/ThemedText";
-import DropDownPicker from "react-native-dropdown-picker";
-import { useController, useForm } from "react-hook-form";
-
-const TextInputWrapper = ({ name, control }) => {
-  const { field } = useController({ control, defaultValue: "", name });
-  return (
-    <TextInput
-      value={field.value}
-      onChangeText={field.onChange}
-      style={styles.input}
-    />
-  );
-};
-
-const DropDownWrapper = ({ choices, control, name }) => {
-  const { field } = useController({ control, defaultValue: "", name });
-
-  const [open, setOpen] = useState(false);
-
-  const [items, setItems] = useState(choices);
-
-  return (
-    <DropDownPicker
-      open={open}
-      value={field.value}
-      items={items}
-      setOpen={setOpen}
-      setValue={(callback: any) => field.onChange(callback())}
-      setItems={setItems}
-    />
-  );
-};
+import { useForm } from "react-hook-form";
+import { TextInputWrapper } from "@/components/form/TextInputWrapper";
+import { SingleSelectWrapper } from "@/components/SingleSelectWrapper";
 
 export default function Index() {
   const { signOut } = useSession();
@@ -63,7 +33,7 @@ export default function Index() {
       style={{ flex: 1, justifyContent: "center", alignItems: "center" }}
     >
       <ThemedText>Submit a Ticket</ThemedText>
-      <DropDownWrapper control={control} choices={choices} name={"label"} />
+      <SingleSelectWrapper control={control} choices={choices} name={"label"} />
       <TextInputWrapper name={"title"} control={control} />
       <TextInputWrapper name={"body"} control={control} />
       <Button
@@ -82,14 +52,3 @@ export default function Index() {
     </ThemedView>
   );
 }
-
-const styles = {
-  input: {
-    height: 40,
-    borderWidth: 1,
-    borderColor: "black",
-    width: 300,
-    padding: 10,
-    margin: 10,
-  },
-};
